+++ app/src/main/java/com/sedsoftware/yaptalker/presentation/extensions/AppCompatActivityExt.kt
@@ -1,21 +1,34 @@
 package com.sedsoftware.yaptalker.presentation.extensions
 
+import android.support.annotation.ColorRes
 import android.support.v7.app.AppCompatActivity
-import android.widget.Toast
-import es.dmoral.toasty.Toasty
+import android.widget.TextView
+import com.androidadvance.topsnackbar.TSnackbar
+import com.sedsoftware.yaptalker.R
 
-fun AppCompatActivity.toastError(message: String) {
-  Toasty.error(this, message, Toast.LENGTH_SHORT, true).show()
+fun AppCompatActivity.snackError(message: String) {
+  showTopSnackbar(message, R.color.snackColorError, R.color.snackColorText)
 }
 
-fun AppCompatActivity.toastSuccess(message: String) {
-  Toasty.success(this, message, Toast.LENGTH_SHORT, true).show()
+fun AppCompatActivity.snackSuccess(message: String) {
+  showTopSnackbar(message, R.color.snackColorSuccess, R.color.snackColorText)
 }
 
-fun AppCompatActivity.toastInfo(message: String) {
-  Toasty.info(this, message, Toast.LENGTH_SHORT, true).show()
+fun AppCompatActivity.snackInfo(message: String) {
+  showTopSnackbar(message, R.color.snackColorInfo, R.color.snackColorText)
 }
 
-fun AppCompatActivity.toastWarning(message: String) {
-  Toasty.warning(this, message, Toast.LENGTH_SHORT, true).show()
+fun AppCompatActivity.snackWarning(message: String) {
+  showTopSnackbar(message, R.color.snackColorWarning, R.color.snackColorText)
+}
+
+fun AppCompatActivity.showTopSnackbar(message: String, @ColorRes bgColor: Int, @ColorRes textColor: Int) {
+  TSnackbar
+    .make(findViewById(R.id.content_container), message, TSnackbar.LENGTH_SHORT)
+    .also { snackbar -> snackbar.view.setBackgroundColor(color(bgColor)) }
+    .also { snackbar ->
+      (snackbar.view.findViewById(com.androidadvance.topsnackbar.R.id.snackbar_text) as? TextView)
+        ?.setTextColor(color(textColor))
+    }
+    .show()
 }