+++ app/src/main/java/com/sedsoftware/yaptalker/presentation/extensions/FragmentExt.kt
@@ -1,21 +1,36 @@
 package com.sedsoftware.yaptalker.presentation.extensions
 
+import android.support.annotation.ColorRes
 import android.support.v4.app.Fragment
-import android.widget.Toast
-import es.dmoral.toasty.Toasty
+import android.widget.TextView
+import com.androidadvance.topsnackbar.TSnackbar
+import com.sedsoftware.yaptalker.R
 
-fun Fragment.toastError(message: String) {
-  context?.let { Toasty.error(it, message, Toast.LENGTH_SHORT, true).show() }
+fun Fragment.snackError(message: String) {
+  showTopSnackbar(message, R.color.snackColorError, R.color.snackColorText)
 }
 
-fun Fragment.toastSuccess(message: String) {
-  context?.let { Toasty.success(it, message, Toast.LENGTH_SHORT, true).show() }
+fun Fragment.snackSuccess(message: String) {
+  showTopSnackbar(message, R.color.snackColorSuccess, R.color.snackColorText)
 }
 
-fun Fragment.toastInfo(message: String) {
-  context?.let { Toasty.info(it, message, Toast.LENGTH_SHORT, true).show() }
+fun Fragment.snackInfo(message: String) {
+  showTopSnackbar(message, R.color.snackColorInfo, R.color.snackColorText)
 }
 
-fun Fragment.toastWarning(message: String) {
-  context?.let { Toasty.warning(it, message, Toast.LENGTH_SHORT, true).show() }
+fun Fragment.snackWarning(message: String) {
+  showTopSnackbar(message, R.color.snackColorWarning, R.color.snackColorText)
+}
+
+fun Fragment.showTopSnackbar(message: String, @ColorRes bgColor: Int, @ColorRes textColor: Int) {
+  activity?.let { activity ->
+    TSnackbar
+      .make(activity.findViewById(R.id.content_container), message, TSnackbar.LENGTH_SHORT)
+      .also { snackbar -> snackbar.view.setBackgroundColor(activity.color(bgColor)) }
+      .also { snackbar ->
+        (snackbar.view.findViewById(com.androidadvance.topsnackbar.R.id.snackbar_text) as? TextView)
+          ?.setTextColor(activity.color(textColor))
+      }
+      .show()
+  }
 }