+++ app/src/main/java/com/sedsoftware/yaptalker/presentation/features/userprofile/UserProfileFragment.kt
@@ -11,8 +11,8 @@ import com.sedsoftware.yaptalker.presentation.base.BaseFragment
 import com.sedsoftware.yaptalker.presentation.base.enums.navigation.NavigationSection
 import com.sedsoftware.yaptalker.presentation.extensions.hideView
 import com.sedsoftware.yaptalker.presentation.extensions.loadFromUrl
+import com.sedsoftware.yaptalker.presentation.extensions.snackError
 import com.sedsoftware.yaptalker.presentation.extensions.stringRes
-import com.sedsoftware.yaptalker.presentation.extensions.toastError
 import com.sedsoftware.yaptalker.presentation.model.base.UserProfileModel
 import kotlinx.android.synthetic.main.fragment_user_profile.*
 import java.util.Locale
@@ -51,7 +51,7 @@ class UserProfileFragment : BaseFragment(), UserProfileView {
   }
 
   override fun showErrorMessage(message: String) {
-    toastError(message)
+    snackError(message)
   }
 
   override fun displayProfile(profile: UserProfileModel) {